Refactoring some more, added a file for minecraft startup
