+++ minecraft/suit/Main.c
@@ -1,38 +1,12 @@
 #include "Utilities/MinecraftClient.h"
+#include "Utilities/MinecraftStartup.h"
+#include "Utilities/Directions.h"
 
 int main() {
 	struct Graph* graph = createGraph();
 
 	//taking out minecraft connectivity for now
-
-	/*
-	if (mcInit() != 0) {
-		printf("mcInit error\n");
-		return 1;
-	}
-
-	if (mcConnect(25566) != 0) {
-		printf("mcConnect  error\n");
-		return 1;
-	}
-
-	mcResetGraph();
-	mcGetNodes(graph);
-
-	int WHITE = 0xFF'FF'FF'FF; // RGBA
-	int GREEN = 0x00'FF'00'FF;
-
-	mcStartUpsertGraph();
-
-
-	for (int i = 0; i < graph->used; i++) {
-		struct Node* node = graph->nodes[i];
-		for (int j = 0; j < node->adjacent; j++) {
-			mcUpsertEdge(node->ID, node->adjacencyArray[j]->ID, WHITE);
-		}
-	}
-	mcStopUpsertGraph();
-	*/
+	//startup(graph);
 
 	FILE* file;
 	fopen_s(&file, "Data Sets/test_points.txt", "r");
