+++ minecraft/suit/Utilities/MinecraftStartup.h
@@ -1,0 +1,9 @@
+#ifndef _MC_STARTUP_H
+#define _MC_STARTUP_H
+
+#include "Directions.h"
+#include "MinecraftClient.h"
+
+void startup(struct Graph* graph);
+
+#endif
